Update crud produk

+++ database/migrations/2024_05_25_085742_create_produks_table.php
@@ -17,7 +17,7 @@ class CreateProduksTable extends Migration
             $table->id();
             $table->string('kode_produk')->unique(); // Membuat kode produk unik dan tidak nullable
             $table->string('nama_produk')->nullable();
-            $table->decimal('harga', 15, 2)->nullable(); // Menyesuaikan format harga dengan presisi yang lebih baik
+            $table->float('harga')->nullable(); // Menyesuaikan format harga dengan presisi yang lebih baik
             $table->integer('stok')->unsigned()->default(0); // Menetapkan nilai default untuk stok
             $table->text('deskripsi')->nullable();
             $table->string('status')->nullable();
